+++ src/mailman/docs/NEWS.rst
@@ -19,6 +19,7 @@ Bugs fixed
 * Incoming messages rejected with defects now include the defects list in the
   status and the rejection and defects list is logged.  (Closes #351)
 * A member's delivery_status is now available via REST.  (Closes #1184)
+* Subscribe and unsubscribe events are now logged.  (Closes #1143)
 
 .. _news-3.3.10: